Move getEnv into its own file
